requirements.txt
streamlit
pandas
numpy
mysql-connector-python
feedparser
beautifulsoup4
openai
requests
